+++ netket/optimizer/qgt/qgt_jacobian_common.py
@@ -67,6 +67,7 @@ def to_shift_offset(diag_shift, diag_scale):
     else:
         return diag_scale, diag_shift / diag_scale
 
+
 @partial(jax.jit, static_argnames="ndims")
 def rescale(centered_oks, offset, *, ndims: int = 1):
     """